* Get rid of the animations when loading the context, they're a little annoying right now. * Fix warnings
svn path=/trunk/extragear/multimedia/amarok/; revision=666946

+++ src/contextview/albumbox.cpp
@@ -82,6 +82,6 @@ void AlbumBox::addAlbumInfo( const QString &pixLocation, const QString &text )
     albumRow->setPos( 0, m_bottom );
     m_bottom += albumRow->bottom();
 
-    setContentRectSize( QSize( boundingRect().width(), m_bottom ) );
+    setContentRectSize( QSize( (int)boundingRect().width(), (int)m_bottom ) );
 }
